Creating an enhancement for DTK consumption.
Enhancements are used in OCP and k8s projects in order to propose some
design to a feature that may involved other projects or teams so we can
get a wider review for it.

This PR can be updated according to comments and then:

If merged, we know that we have a well defined design that was approved
and is ready for implementation.

If closed, then we have an historical documentation oh why this proposal
was rejected.