Fix duplicate namespace inspects for DSs in must-gather (#72)
This change removes duplicate execution of `oc adm inspect daemonsets`
statements, when fetching the namespaces to loop for by getting the
DaemonSets of all namespaces labelled as KMM DSs. There could be 2
KMM DaemonSets per namespace, i.e. the module-loader and the
device-plugin one. As a result, the namespaces array for which to loop
included duplicates.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

Signed-off-by: Michail Resvanis <mresvani@redhat.com>