add code/dockerfile for building the secureboot signing job image (#101)
This commit adds a binary (signimage.go) and an associated Dockerfile
that together will provide the image for use in the signing job.

The dockerfile compiles signimage, but it also retrieves the sign-file
binary from the kernel-devel package. This binary is the one provided by
the kernel developers for signing kmods, and while its oepration is not
super-complex and could be replaced, we use it to ensure maximum
compatability and future-proofing.

The signimage.go binary takes an image name, pulls it down, loops through
each file in each layer and if its a kernel modules extracts it to a
temporary directory, and runs sign-file on it. All signed kmods are then
added back as a (single) new layer, and the new image pushed back to the
registry.

To support this extensive functionality is also added to internal/registry/registry.go