Fixing a bug in the 'sign' UT. (#163)
The code is using the same global module object for all tests, which
means that if one of the specs change the module, it affects all other
tests.

Each spec should be tested with a clean module.

How did it impact the code before the change:

Running
```
make unit-test TEST=github.com/kubernetes-sigs/kernel-module-management/internal/sign/job
```
was working only because all the tests were running one after the other,
once we focused on the `should set correct kmod-signer flags` spec only
it was failing because it was expecting some state which was modified by
another spec that run before it.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>

Upstream-Commit: 2642a7db8531314c6ae4f2504cd1e4c0dc69d3aa