Add the kernel - ClusterClaim reconciler (#152)
This reconciler collects all kernel versions running in the cluster and
populates a ClusterClaim with them.
To minimize updates to the ClusterClaim (which triggers a sync back to
the ManagedCluster), kernel versions are sorted alphabetically and
separated by a new line.
This controller is not started by any manager yet because we still have
to determine if we add it to the mainline KMM or if we create a
dedicated image for Spoke clusters.

Upstream-Commit: ac9163fd083aa48092b1e5c15650a87392fbdf69

Adjust ClusterClaim permissions to enable creation and caching (#191)

The NodeKernelClusterClaim reconciler requires permissions to create,
patch and also leverage client caching on ClusterClaim resources. This
change adds permissions to [create,get,list,watch] ClusterClaim
resources, but only [patch,update,delete] on the specific ClusterClaim
the reconciler itself has created.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

Upstream-Commit: a88bef3b712b6fbd784079d5a995c468a14c2480