refactoring GetHeaderDataFromLayer function
GetHeaderDataFromLayer function is also used to verify that
a file(header) exists in an image(layer). it check for header
existence and reads its input into a buffer. Since this function
can be used to verify a kernel module file existence, it can be problematic
to read a whole file into a memory, since kernel module file can be a big one.
This PR split this function into an inner implementation that checks header
existence and returns a io.Reader to it, and the upper implementation that
is actually using the io.Reader to read the data