Fix module reconciler not being able to watch for ConfigMaps error (#179) (#225)
This fix adds the watch permission for ConfigMaps to the Module
reconciler, in order to be able to successfully client.Get() the build
Dockerfile ConfigMap. This client.Get() fails because the
contoller-runtime client reads from its cache, instead of directly from
the API server. The client needs the watch permission, in order to be
able to build its cache.

- https://github.com/kubernetes-sigs/controller-runtime/issues/1156

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

Upstream-Commit: dedee81d338667c32f9bee2b26fc439ca022c391

Co-authored-by: Michail Resvanis <mresvani@redhat.com>