Adding PreflightOCP controller
Preflight OCP controller is responsible for initiating the preflight flow
in OCP cluster. The difference between PreflightOCP and Preflight is that
the input to PreflightOCP is an OCP release image and not kernel version.
In order to utilize the already existing Preflight controller, the flow
of the PreflightOCP controller is as following:
1) extract DTK, OS version and kernel version from the input OCP image
2) update the kernel dtk mapping with the extracted data
3) create a new Preflight CR with the extracted kernel version
4) monitor status updates of the Preflight and update Preflight OCP
   Status with the data taken from Preflight CR Status

This way we are not interfering with the upstream flow of the Preflight, and
the PreflightOCP controller is just an upper layer that prepares data for the
Preflight flow, and then just copies its statuses