Remove unnecessary permissions on Module resources
KMM does not create nor delete Module Custom Resources.

Upstream KMM does not require the update permission on Module finalizers
neither. But one of the default enabled Admission plug-ins in OpenShift
is the OwnerReferencesPermissionEnforcement, which protects the access
to metadata.ownerReferences[x].blockOwnerDeletion of an object, so that
only users with update permission to the finalizers subresource of the
referenced owner can change it. Since the KMM Module reconciler needs to
set a Module owner reference with the blockOwnerDeletion enabled on the
resources it's managing, it requires the update Module finalizers
permission.

- https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#ownerreferencespermissionenforcement
- https://access.redhat.com/documentation/en-us/openshift_container_platform/4.11/html/architecture/admission-plug-ins

Signed-off-by: Michail Resvanis <mresvani@redhat.com>