Add the ManagedClusterModule CRD and reconciler (#153) (#245)
The cluster-scoped ManagedClusterModule CRD is a new Module wrapper, that
features a selector for ManagedClusters to target specific Spoke clusters.

The respective reconciler, hosted on the Hub cluster, creates a ManifestWork
in the corresponding namespace, in order to deploy a Module in the Spoke
cluster. When crafting the ManifestWork, the reconciler removes any build
or sign sections.

Notes
-----
- KMM already supports building and signing kernel modules via the
  Module CRD. This reconciler will use the same features, in order to
  build and sign the kernel modules only on the Hub. This feature will
  be implemented in a future PR.

- The ManagedClusterModuleStatus is also going to be implemented in a
  future PR.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

Upstream-Commit: b469c654418be5e278610cc25a5b7ad9ef7302ab

Add empty ManagedClusterModuleStatus to resolve operator-sdk error (#169)

operator-sdk complains about the missing status of a CRD. This change
adds an empty ManagedClusterModuleStatus to the ManagedClusterModule
CRD, which will be implemented in a future PR.

- https://github.com/operator-framework/operator-sdk/blob/a1b976e7d40e4c8c5c3800283de0c0905c06bd56/internal/generate/clusterserviceversion/bases/definitions/crd.go#L109

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

Signed-off-by: Michail Resvanis <mresvani@redhat.com>
Co-authored-by: Michail Resvanis <mresvani@redhat.com>