Adding TLS options to the sign logic. (#272)
When checking if the `signed` image already exist or when pulling the
image to sign, we should be able to set some TLS options.

This flow follows the same logic we have for builds.

In additions, this commit also update the type of the
`ModuleLoaderContainerSpec.RegistryTLS` API so we use the default values
when not specified.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>
Co-authored-by: Yoni Bettan <yonibettan@gmail.com>