Installing `operator-skd` via Makefile. (#219)
To make sure u/s and m/s are always using the same version without the
need of manually installing and maintaining `operator-sdk` with the same
version on u/s infra and on m/s infra.

In addition to that, also updating all bundle files according to the new
`operator-sdk` version. CI will enforce it from now on.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>