Make buidconfig build the intermediate image if sign is defined
If both build and sign are defined build (buildconfig) should build an intermediate image
and sign then detects the final image doesn't exist so uses the intermediate image to create it

Due to a bug buildconfig was creting the final image directly even when sign was defined, so when
build completed KMM detected the final image existed and did not hand over to the sign job to sign the kmod.
This meant in the case build AND sign were required sign was never run.

This commit fixes that issue by forcing build config to create the correct (i.e. intermediate) image
if sign is defined