Add CA ConfigMaps in each namespace for sign jobs
Signing Jobs can be created in any namespace, while the ConfigMaps containing the OpenShift CA
certificates from the bundle are only present in the operator's namespace.
In each namespace containing a Module or targeted by a ManagedClusterModule, create two ConfigMaps
for each CA certificate type. Those ConfigMaps are populated by other OpenShift controllers.
When creating a signing Job, look for those ConfigMaps in the Job's namespace and mount them
accordingly.
Set the ConfigMaps owner references appropriately so that they are deleted when all Modules /
ManagedClusterModules that might need it are deleted.