Fixes #359 (#364)
(this fixes a small discrepancy between upstream and midstream infavour of upstream)

log error if sign-file fails (#237)

the error from the sign-file binary was getting lost because it wasn't
returned through the stack call correctly. This commit fixes that by
passing the message up.

Also remove a couple of commented out code lines.