Adding module version labels definitions (#330)
This PR add the following 3 labels, that will be used during ordered
update process:
1) kmm.node.kubernetes.io/version-label-module_<namespace>_<name>
2) kmm.node.kubernetes.io/version-label-module-loader_<namespace>_<name>
3) kmm.node.kubernetes.io/version-label-device-plugin_<namespace>_<name>

The first label is set by the user and signifies on which node which version
should be run. The second and the third labels are managed by KMMO and are used as
node selectors for module loader and device plugin damoensets accordingly.
This PR also include utils functions for generating the full labels' names and
for extracting namespace and name data from labels

Upstream-Commit: a5ebe9b8564ecdcb3767211bab2ab82d4681094a