Adding a `prow` script to test a basic deployment. (#353)
This test will make sure that steps added to the `e2e` workflow are not
braking our deployment flow mentioned in
kernel-module-management#getting-started or by using `make deploy` which
are basically the same.

A good example for it is to add a validation web-hook that requires the
installation of `cert-manager` which should also be installed outside of
the CI environment.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>

Upstream-Commit: 91e4bca6ddb1eecda65522368b8ae89618129c82