Adding api for ordered multiple kernel modules (#380)
Adding list entry under Modprobe spec that defines
the sequence of loading multiple independent kernel modules
that are stored in the Module image. The entry is a list with
the modules that depends on all other modules at the head of the list.
Example: if moduleA depends on first loading moduleB, and moduleB depends on first loading moduleC
then the list will look:
ModulesLoadingOrder:
- moduleA
- moduleB
- moduleC
moduleC shoudl be loaded first, then module B, and then module C.
moduleA should be set in the ModuleName entry of the ModprobeSpec

Co-Authored-By: Quentin Barrand <quentin@quba.fr>
Upstream-Commit: e5144dd57d0821b2facf484bb5e18c44e2687a55