Adding the namespace to the `<module>.ready` label. (#526)
This is handy when there are multiple modules on each node and make it
easier to determine where is each module.

Also, solving a race condition in a new pod was created while the old
pod of a `moduleLoader` is being terminated which resulted in the
`<module>.ready` label to be deleted.


Upstream-Commit: bc488304f926742c14b0837563ef63acfef2b158

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>
Co-authored-by: Yoni Bettan <yonibettan@gmail.com>