Loading multiple independent kernel modules in one KMM Module (#381) (#556)
This PR allows defining the load order of multiple independent
kernel modules in one KMM Module. The order is defined as a list
in the ModulesLoadingOrder field of the ModprobeSpec.The first
module will be loaded last, the last one in the list will be loaded
first. It is implement by using DownwardAPI: module loader pod template
contains is added an annotation that contains the
/etc/modprobe.d/softdep format, and then this annotation is added
as a file to module loader's pod filesystem using Downward API