Upgrade support between v1.02 and v1.1 (#639)
Currently in v1.1 the naming scheme for ModuleLoader and DevicePlugin
daemonsets is different the what we had in v1.0.2. We are using hashings
and the names are deterministic. This causes issues during upgrade,
since the upgraded controller will try to create new daemonsets (new
names). The old daemonsets can not be deleted, since it will cause
the unloading of kernel modules. Since there is no way to change/delete
lifecycle hooks of the running pod, we need to support the naming scheme
of v1.0.2
The changes in PR:
1. get the runnings Daemonsets of the module in the earlier stage of the
   reconciliation loop
2. during ModuleLoader or DevicePlugin handling, check if the
   appropriate daemonset already exists based on the kernel-version and
   version label of daemonset
3. if now daemnonset is found, then create a new one using GenerateName
   option
4. removing the fix for pods being stuck in Terminating state for ~30
   seconds, since it changes the Container spec of the Daemonset, which
   means that during upgrade the DaemonSet will delete the pod and then
   restart it, which may cause unloading of kernel module
5. unit test fixes

Upstream-Commit: a0d6cb90271543682c8d556b3e02df721efe20bf