Fixing ModuleLoader recreation during Node reboot (#464) (#655)
Bug sequence as following:
1) Module deployed, with selector targeting only one node
2) ModuleLoader DS created and pod is running
3) Targeted Node rebooted
4) Once Node becomes NotReady, DS is destroyed
5) Node becomes Ready again, but no new ModuleLoader DS is created (bug)

The reason for bug is that once node is in the cluster, all the event
for node are Update events. Since the labels do not change between the
time that node was NotReady to NodeReady, the events are filtered our
and the reconciliation loop is not started

Solution: add another predicate to the filtering of the node event
for module reconciler. This predicate will return true in case the
Node's condition for Ready type has chnaged and the new condition status
is True