Adding NotFound error to the GetMLD API (#502)
In case kernel mapping for the kernel version is not found, GetModuleLoaderDataForKernel
API returns an indistinguishable error, just like in case there was some
internal error during processing.
This PR introducing a new propriaetory error: ErrNotFound. In case
kernel mapping for a specific kernel does not exists, this error will be
returned. As a consequence, the calling code will be able to act
seperatly on internal error, or on KernelMapping not present, which may
not be an error at all

Upstream-Commit: bb43458942a744449c94232b0f664e2c566fd28a