Fixing Module with device-plugin finalization deadlock (#590) (#806)
Currently, when the module is deleted, the finalization function is
called. This function basically waits till all the kernel modules has
been unloaded, and only then removes the finalizer from Module.
In case device-plugin is deployed, its Daemonset won't be deleted (since
Module is not deleted due to finalizer), and the kernel modules cannot
be unloaded due to device plugin running
Solution: in case of Module being deleted, handle device-plugin deletion
in the module-reconciler specifically, and not wait for it to be deleted
based on ownerReference