Fix prepareSchedulingData loop variable capture (#591)
Sending the *v1.Node loop variable from prepareSchedulingData to
prepareNodeSchedulingData led to an invalid scheduling data map in which
all values' node field was pointing to the same v1.Node object - the
latest item in the targetedNodes slice.
Send the v1.Node to prepareNodeSchedulingData instead of a pointer to
solve this problem.

Upstream-Commit: c3eafe349de684497d47a09f96a8b09b295937b0