Adding support for on-prem OCP clusters to day1 scenario (#1066)
On-prem OCP cluster, installed via assisted-installed, has an
additional service that is resposible for configuring DNS on host.
Therefore usual day1 service, that is dependent only on NM service, will
be unable to pull images prior to DNS configuration service being
completed. Since systems service do not support "if else" syntax
we will define 2 pull-kernel-module-image services:
1) service that is supposed to be run on usual cluster will be executed
   only in case /etc/systemd/system/on-prem-resolv-prepender.service
   file is missing from the host
2) service that is supposed to be run on on-prem clusters will run only
   in case /etc/systemd/system/on-prem-resolv-prepender.service is
   present on the host, and will depend both on NM service and on
   NetworkManager-dispatcher service

Since NetworkManager-dispatcher service is a dbus service, the second
service should verify if it has completed successfully before running
the main executable. This is acomplished by a dedicated script that wait
on NetworkManager-dispatcher service and runs as a PresStart script