Changing the `ImagePullPolicy` for the operator and the webhook. (#1167)
For the KMM operator, the `ImagePullPolicy` was set to `Always` for
development reasons but it is in fact not the best practice, therefore,
it was removed to allowed `kubelet` to have its default pull policy.

The webhook-server's pull policy was also removed to use the default
behavior to stay in sync with the KMM operator it is deployed with.

In a nutshell, the default policy is `Always` if the `latest` tag
is used and `IfNotPresent` otherwise.

This change is expected to reduce the networking bandwidth.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>