Changing Node interface and updating NMC controller
This PR is a preparation for the fixes needed for cluster upgrade KMM
implementation. We are adding a new function to the node interface,
that determines whether a node has been rebooted while a kernel module
was loaded. This PR also updates NMC controller to use this API and
refactors the unit-test