Add support for user defined tolerances to modules (#1257)
There is a requirement to run selected pods on a cordoned node.
One such example is device driver upgrade. The procedure involves
the node to be cordoned. At the same time, there should be a way
to run house keeping pods that carry out the driver upgrades. These
pods should be schedulable.
This PR adds support to carry the tolerances for the pods that carry
out the house keeping operations. These user defined tolerances will
match the taint that gets added to the nodes.

ModuleSpec will be used to carry the tolerance to the pods which will
be used during pod creation.