Adding a `.` as a valid character for `Module`s' name. (#1294)
When a `Module` is deleted, the `NMC` controller is removing the labels
for this `Module` in the `NMC` object.

As long as those labels for a specific module exists, the module's
finalizer will not be removed upon module deletion and the module will
keep hanging forever if deleted.

When a module is created with a `.` in its name, the created label for
`NMC` doesn't match the expected regexp for this label and, therefore, not
removed upon module's deletion.

Adding the `.` will allow the label to match the regexp, be removed upon
deletion and deleting the module gracefully.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>