Fixing toleration implementation (#1338)
In current implementation, tolerations are stored in the ModuleConfig of
the NMC. That means that on each update of Tolerations, kernel module
will be unloaded/loaded, which is undesirable.
In this PR, the tolerations are moved into ModuleItem, along with Name,
Namespace,Service Account and PullSecret, since it does not affect the
kernel module code, but just the way the worker pod is configured to
run. This way, tolerations can be changed in whatever way possible,
without affecting the loaded kernel modules

Co-authored-by: Yevgeny Shnaidman <yshnaidm@redhat.com>