Removing the `Generation` field from the `ModuleImagesConfigSpec`. (#1341)
At first, we thought we will need a field to let the `MIC` future
controller know that it needs to retry pulling an image that didn't
exist, for example if the user has forgot to push the image.

We decided to just set the future pull pod with `restartPolicy=OnFailure`
instead that will constantly retry to run the pod when it fails with an
exp waiting time before retries and the max waiting time being 5m. If we
wish to manually retry the pod without waiting the current waiting time, we
can just delete the pull pod and the `MIC` controller will re-create it
instantly.

It removed the need for maintaining a `Generation` field separately.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>