Use the correct pull secret for image pulls in builds
Currently, when a build attempts to
pull an image from a private registry,
it fails due to the absence of the appropriate pull secret.
This commit ensures that the module's
pull secret is included in the build's YAML,
allowing it to authenticate to the private
registries and pull images from there.