Un-exporting some of `WorkerPodManager` interface env variables.
When this interface was transitioned to the `pod` package, it was copied
as is (as much as possible) and some variables had to be exported to be
used as before.

Once the transition done, it was simpler to review some of those new
exported env variables and convert them to usefull methods inside the
`WorkerPodManager` interface.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>