Allow KMM to be installed on worker nodes if there is no better choice.
Some installations of k8s doesn't allow deploying workloads on the
control-plane nodes or the control-plane isn't accessible at all.

In these installs, the worker nodes do not have the
node-role.kubernetes.io/control-plane or node-role.kubernetes.io/master
labels. Because the worker nodes do not have these labels the Kernel Module
Management operator will not run out of the box.

This commit changes the nodeAffinity to
preferredDuringSchedulingIgnoredDuringExecution to solve this issue.

Signed-off-by: brucejcong
Co-authored-by: Yoni Bettan <yonibettan@gmail.com>