Updated the MCM controller to work with the new MIC implementation.
The MCM controller doesn't have to create build/sign builds manually
anymore, all it needs to do is to create 1 MIC object in the cluster per
MCM object per managed cluster.

If all images in an MIC are ready, then the controller can proceed with
creating the manifestWork because it guarantees that the targeted managed
cluster can pull all kmod images successfully.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>