Adding manager and maker/signer to buildsign package
With the introduction of the MBSC controller, the flow is changing and
there is no need for 2 different managers. There is no need to decide if
the build/sign should be executed, since that decision is taken by the MIC controller.
The new package must provide only the following capabilities:
1. Get the running OCP Builds (build or sign), in order for MBSC controller to
   update its status
2. Sync the current build/sign OCP Builds based on action parameter
3. Delete build/sign successfull OCP Builds based on the action parameter

The following files are added:
1. manager - it will decide, based on the action field, which action
   should be taken: check OCP Build, create OCP Build etc'. It will use a dedicated
maker/signer interfaces when creating/deleting relevant pod. In addition
it will return the status of the running build/sign OCP Builds
2. maker - will be responsible for creating the "Build" OCP Build. Copied almost
   as is from the current build package
3. signer - will be responsible for creating the "Sign" OCP Build. Copied almost
   as is from the current sign package