Adding SkipWaitMissingImage field to MIC Spec
Currently, in case image is missing and Build and Sign sections are not
defined, MIC controller will keep the image puller pod running and
waiting till the images appears. This means that pushing the mising
image into the repo will cause the image puller to succeed and the KMM
flow will continue running without any need for additional user action.
The new field, if set to true, will cause the MIC to report the image as
missing and abort the image puller pod (in case Build and Sign sections
are not defined). Currently it will be used in the Preflight flow