Reducing the `buildsign`'s porting layer.
Instead of making the `Manger` a generic interface consuming a porting
layer of its implementation, we can lower down the porting layer in the
stack and make the `ResourceManager` the generic interface consuming a
porting layer of its implementation to be the porting layer instead.

This makes more code generic and reduce the size of the porting layer.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>