Removing the base `manager_config_patch.yaml` file.
As part of 7b0f9b476b3ed9378e9c465bd26898044eec0544, 2 patches were
created for the config, one for `manager` and one for `manager-hub`
patching the same controller argument.

Both of those patches, are pathing a non-valid patch applied by the
`manager-base` layer, which is not valid but doesn't flag an error as it
is being overwritten by valid values in upper kustomization layers such
as `manager` and `manager-hub`.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>