Considering the `New` build status when updating the MBSC status.
The MBSC status is updated based on the build statuses. We were missing
the `New` status which would log some errors until the build phase
matures to a later phase.

This commit isn't changing the functionality but just making sure we
don't log errors due to a temporary, end expected, phase.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>