Introducing mcfg (machine config) package
Currently pkg/mcproduce includes a code that creates the MachineConfig
for Day1. we are going to have a controller that will also need to
create/update a MachineConfig, specifically ignition part. This commit
introduces a mcfg package, that contains API that creates an ignition
json, and will be used by the mcproducer and the future controller.
Changes:
1. templates and scripts involved in the ignition are moved to mcfg
   package
2. an API for generating the Ignition (using scripts and templates) is
   introduced
3. in pkg/mcproduces, the machine config package is changed to just
   imprt the ignition created by the mcfg API