Adding documentation for the versioned ready label on the nodes.
Users need an official API to determine if their kmod was upgraded to
the newer version successfully when using the ordered-upgrade flow.

Before this commit, we only had a label describing if a specific kmod is
ready to be used or not but the same label was used for all versions.

To prevent users from watching the ready label being removed and then
re-appearing for the new kmod, we decided to add a new label with the
kmod version in it.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>