Day1 utility isn't working without firmware.
When creating the worker pod in the day1 utility, we always mount the
firmware directory from the host but we only create the directory when
there are some firmware files to apply which lead to the container
creation failure.

This commit is fixing this issue.

I also changed the logs to be clearer when debugging.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>