Fixing KMM deployment for clusters under RedHat SRE (ROSA)
Openshift cluster that are managed by Redhat SRE, do not allow(restrict)
user from changing ignitions etc'. It means that the MachineConfig CRD
is not deployed on the cluster, which causes the BMC controller to
constantly fail, since it is watching for that object.
The fix is to check the presence of the MachineConfig CRD, and to
kick-off the BMC controller only in case it is present